chore: refactor pacakge
