+++ app/src/main/java/com/teamopendata/mindcareapp/common/model/entity/Task.java
@@ -1,7 +1,6 @@
-package com.teamopendata.mindcareapp.model.entity;
+package com.teamopendata.mindcareapp.common.model.entity;
 
 import androidx.annotation.NonNull;
-import androidx.room.ColumnInfo;
 import androidx.room.Entity;
 import androidx.room.Ignore;
 import androidx.room.PrimaryKey;
