+++ app/src/main/java/com/teamopendata/mindcareapp/converters/Converters.java
@@ -2,20 +2,18 @@ package com.teamopendata.mindcareapp.converters;
 
 import androidx.room.TypeConverter;
 
-import com.google.android.gms.tasks.Tasks;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
-import com.teamopendata.mindcareapp.model.entity.Task;
-import com.teamopendata.mindcareapp.util.Utils;
+import com.teamopendata.mindcareapp.common.model.entity.Task;
+import com.teamopendata.mindcareapp.common.Utils;
 
 import java.lang.reflect.Type;
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
 
 public class Converters {
     @TypeConverter
-    public String fromTasksList(List<Task> list) {
+    public String fromTaskList(List<Task> list) {
         if (list == null) return (null);
         Type type = new TypeToken<List<Task>>() {
         }.getType();
@@ -23,7 +21,7 @@ public class Converters {
     }
 
     @TypeConverter
-    public List<Task> toTasksList(String value) {
+    public List<Task> toTaskList(String value) {
         if (value == null) return (null);
         Type type = new TypeToken<List<Task>>() {
         }.getType();
