+++ app/src/main/java/com/teamopendata/mindcareapp/ui/records/listener/OnAddEditRecordListener.java
@@ -1,6 +1,6 @@
 package com.teamopendata.mindcareapp.ui.records.listener;
 
-import com.teamopendata.mindcareapp.model.entity.Record;
+import com.teamopendata.mindcareapp.common.model.entity.Record;
 
 public interface OnAddEditRecordListener {
     void onAddEditRecordButtonClick(Record record);
